Add docs and logs
